+++ app/src/main/java/ru/gb/student/gb_popular_lib/App.kt
@@ -1,0 +1,12 @@
+package ru.gb.student.gb_popular_lib
+
+import android.app.Application
+import android.content.Context
+import ru.gb.student.gb_popular_lib.data.FakeUsersRepoImpl
+import ru.gb.student.gb_popular_lib.domain.UsersRepo
+
+class App : Application() {
+    val usersRepo: UsersRepo by lazy { FakeUsersRepoImpl() }
+}
+
+val Context.app:App get() = applicationContext as App
