+++ app/src/main/java/ru/gb/student/gb_popular_lib/domain/UserEntity.kt
@@ -1,4 +1,4 @@
-package ru.gb.student.gb_popular_lib
+package ru.gb.student.gb_popular_lib.domain
 
 data class UserEntity (
     val login: String,
