+++ app/src/main/java/ru/gb/student/gb_popular_lib/domain/UsersRepo.kt
@@ -1,4 +1,4 @@
-package ru.gb.student.gb_popular_lib
+package ru.gb.student.gb_popular_lib.domain
 
 interface UsersRepo {
     fun getUsers(
