+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".MainActivity">
+    tools:context=".ui.users.MainActivity">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/refresh_button"
